Major rework for v0.2.0
The 0.1.x wrappers do not work (Arc, Mutex, Option, Rc) under any
but the most cursory testing. The issue is the shortcut of using Deref
and DerefMut to preserve the builder interface; this can be done
differently (and now is).

Add persian_rug support via a feature

GeneratorWithPersianRug has additional restrictions:
- No customisation of other GeneratableWithPersianRug items.
- Types required on generator_with_persian_rug expressions.

The other cases can avoid limitations because they do not need to hold
state (these are the only two ways of invoking a nested
GeneratorWithPersianRug). Since they don't hold state, we don't have
to be able to name their type.

Implement GeneratorWithPersianRug for all the built-in generators,
including lambdas.