Work in progress
Implement GeneratorWithPersianRug with the current set of restrictions:
- No customisation of other GeneratableWithPersianRug items.
- Types required on generator_with_persian_rug expressions.

The other cases can avoid limitations because they do not need to hold
state (these are the only two ways of invoking a nested
GeneratorWithPersianRug). Since they don't hold state, we don't have
to be able to name their type.

The lava-api-mock compiled for the first time.