Make the tests trigger type inference problems for realistic closures
This affects default_with_persian_rug and sequence_with_persian_rug.

The fix is to assign the closures to typed variables, so that the
compiler can accurately infer the parameter types; it's not possible
for the user to do this in their lambda, because the actual types are
unknown to them.